requirements.txt
pyzbar~=0.1.9
Pillow~=9.4.0
